increase the size of fapi structures to support 8 antennas

+++ nfapi/open-nFAPI/nfapi/src/nfapi_p7.c
@@ -2533,7 +2533,7 @@ int pack_nr_srs_normalized_channel_iq_matrix(void *pMessageBuf, void *pPackedBuf
     return 0;
   }
 
-  uint16_t channel_matrix_size = nr_srs_normalized_channel_iq_matrix->num_prgs
+  uint32_t channel_matrix_size = nr_srs_normalized_channel_iq_matrix->num_prgs
                                  * nr_srs_normalized_channel_iq_matrix->num_ue_srs_ports
                                  * nr_srs_normalized_channel_iq_matrix->num_gnb_antenna_elements;
   // if (nr_srs_normalized_channel_iq_matrix->prg_size == 0){
@@ -2549,7 +2549,7 @@ int pack_nr_srs_normalized_channel_iq_matrix(void *pMessageBuf, void *pPackedBuf
     channel_matrix_size <<= 2;
   }
 
-  for (int i = 0; i < channel_matrix_size; i++) {
+  for (size_t i = 0; i < channel_matrix_size; i++) {
     if (!push8(nr_srs_normalized_channel_iq_matrix->channel_matrix[i], &pWritePackedMessage, end)) {
       return 0;
     }
@@ -4413,7 +4413,7 @@ int unpack_nr_srs_normalized_channel_iq_matrix(void *pMessageBuf,
     return -1;
   }
 
-  uint16_t channel_matrix_size = nr_srs_normalized_channel_iq_matrix->num_prgs
+  uint32_t channel_matrix_size = nr_srs_normalized_channel_iq_matrix->num_prgs
                                  * nr_srs_normalized_channel_iq_matrix->num_ue_srs_ports
                                  * nr_srs_normalized_channel_iq_matrix->num_gnb_antenna_elements;
 
@@ -4429,7 +4429,7 @@ int unpack_nr_srs_normalized_channel_iq_matrix(void *pMessageBuf,
     channel_matrix_size <<= 2;
   }
 
-  for (int i = 0; i < channel_matrix_size; i++) {
+  for (size_t i = 0; i < channel_matrix_size; i++) {
     if (!pull8(&pReadPackedMessage, &nr_srs_normalized_channel_iq_matrix->channel_matrix[i], end)) {
       return 0;
     }
